Support building up topology using containernet
This extension using containernet to support docker container as mininet
host:
https://github.com/containernet/containernet

To use this extension, you can add a new option in the `General` section
of the `g2.conf` configuration file:

    [General]
    ...
    # <host_name>, <container_name>, <cls>; ...
    host_info = rc, rucio, DynamicDocker ; xrd1, N/A, DynamicDocker

It will allow you to use some specific mininet `Host` subclass to class
`<cls>` to construct the host node. `<cls>` MUST inherit `Host`.

The `DynamicDocker` that we provide can wrap a running docker container
as a mininet host.

Signed-off-by: jensenzhang <hack@jensen-zhang.site>